+++ backend/internal/domain/event.go
@@ -8,12 +8,12 @@ type Event struct {
 	ID              *string  `json:"_id,omitempty"`
 	Title           string   `json:"title"`
 	Description     *string  `json:"description"`
-	StartAt         int      `json:"startAt"`
-	EndAt           int      `json:"endAt"`
+	StartAt         int64    `json:"startAt"`
+	EndAt           int64    `json:"endAt"`
 	RoomID          *string  `json:"roomId"`
 	ParticipantsIDs []string `json:"participantsIDs"`
 	Notes           *string  `json:"notes"`
-	RemindAt        int      `json:"remindAt"`
+	RemindAt        int64    `json:"remindAt"`
 	IsDelete        bool     `json:"isDelete"`
 	CreatedAt       int64    `json:"createdAt"`
 	UpdatedAt       int64    `json:"updatedAt"`
